requirements.txt
pandas
pandas-ta
numpy>=2.2.6
python-dotenv==1.0.1
requests==2.31.0
pymongo==4.6.1
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
scikit-learn
websockets==12.0
